[DIDA] 커뮤니티 > 시끌벅적 게시판 carousel ui

+++ feature/home/src/main/java/com/dida/home/adapter/HotsAdapter.kt
@@ -32,7 +32,7 @@ class HotsAdapter(
         holder.bind(getItem(position))
     }
 
-    override fun getItemId(position: Int): Long = getItem(position).cardId.toLong() * -1
+    override fun getItemId(position: Int): Long = getItem(position).cardId * -1
 
     override fun getItemViewType(position: Int): Int = R.layout.holder_hots
